tests/wpt/web-platform-tests/webusb/usb-supported-by-feature-policy.html

<!DOCTYPE html>
<title>Test that usb is advertised in the feature list</title>
<link rel="help" href="https://w3c.github.io/webappsec-feature-policy/#dom-featurepolicy-features">
<link rel="help" href="https://wicg.github.io/webusb/#feature-policy">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(() => {
    assert_in_array('usb', document.featurePolicy.features());
}, 'document.featurePolicy.features should advertise usb.');
</script>
